backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.9.0
gunicorn>=22.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
